Create cee_uuurds.h
I have made this file publicly available in the best interests of the public at large. Programmers using the Generic Language may use this as a keyword reference. For example, If you are programming with Emojis or Mandarin, this may be the only way to obtain the programming keywords so that you may program in the given language.

It has been 10 years since I gave up programming in English. The source code of the Generic Language is all in Inglish!

The keywords in the compiler are defined as AVL Sets. There are 16 supported languages (if you include Emojis as a language).

The programming public can assist itself by updating this file to support your favorite language. If someone adds a language to this file they can contact me at benedict@gigasopht.net and I will incorporate the changes into the compiler. That way, the world has an international standard for compiler keywords.